+++ crates/public/bunsen/src/kits/speech/whisper/pretrained/mod.rs
@@ -9,6 +9,7 @@ pub use pytorch_utils::*;
 
 #[cfg(all(feature = "whisper-weights", feature = "store_pytorch"))]
 mod load;
-
 #[cfg(all(feature = "whisper-weights", feature = "store_pytorch"))]
 pub use bunsen_bundled_whisper as bundled;
+#[cfg(all(feature = "whisper-weights", feature = "store_pytorch"))]
+pub use load::bundled_vocabulary;